+++ app/decorators_v2/nimble/classes/berserker_decorator.rb
@@ -5,6 +5,11 @@ module Nimble
     class BerserkerDecorator < ApplicationDecoratorV2
       def call(result:)
         @result = result
+        @result['keys'] = %w[str dex]
+        @result['hit_die'] = 12
+        @result['saves'] = %w[str int]
+        @result['armors'] = []
+        @result['weapons'] = %w[melee-str range-str]
         @result
       end
     end