+++ app/serializers/nimble/character_serializer.rb
@@ -5,11 +5,12 @@ module Nimble
     include Deps[cache: 'cache.avatars']
 
     attributes :provider, :id, :name, :created_at, :updated_at, :avatar, :level, :guide_step, :ancestry, :main_class, :skills,
-               :skill_points, :modified_abilities, :initiative, :size, :abilities
+               :skill_points, :modified_abilities, :initiative, :size, :abilities, :armor, :speed, :health, :wounds_max,
+               :wounds_spent
 
-    delegate :skills, :modified_abilities, :initiative, :size, to: :decorator
+    delegate :skills, :modified_abilities, :initiative, :size, :armor, :speed, :wounds_max, to: :decorator
     delegate :data, to: :object
-    delegate :level, :guide_step, :ancestry, :main_class, :skill_points, :abilities, to: :data
+    delegate :level, :guide_step, :ancestry, :main_class, :skill_points, :abilities, :health, :wounds_spent, to: :data
 
     def provider
       'nimble'